+++ packages/elevenlabs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/elevenlabs
 
+## 1.0.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [e7fcc86]
+  - @ai-sdk/provider-utils@3.0.0-beta.7
+
 ## 1.0.0-beta.7
 
 ### Patch Changes