+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/rsc
 
+## 1.0.0-beta.30
+
+### Patch Changes
+
+- ai@5.0.0-beta.30
+
+## 1.0.0-beta.29
+
+### Patch Changes
+
+- Updated dependencies [e7fcc86]
+- Updated dependencies [d92b9a8]
+  - @ai-sdk/provider-utils@3.0.0-beta.7
+  - ai@5.0.0-beta.29
+
 ## 1.0.0-beta.28
 
 ### Patch Changes